GeometryMode - fix truncation errors. #rb Jimmy.Andrews #preflight 63861837170bc34a93b16d23
[CL 23306039 by David Hill in ue5-main branch]

+++ Engine/Plugins/Editor/GeometryMode/Source/GeometryMode/Private/EditorGeometry.cpp
@@ -197,7 +197,7 @@ bool FGeomEdge::IsSameEdge( const FGeomEdge& InEdge ) const
 FVector FGeomEdge::GetWidgetLocation()
 {
 	FVector dir = ((FVector)GetParentObject()->VertexPool[ VertexIndices[1] ] - (FVector)GetParentObject()->VertexPool[ VertexIndices[0] ]);
-	const float dist = dir.Size() / 2;
+	const double dist = dir.Size() / 2;
 	dir.Normalize();
 	const FVector loc = (FVector)GetParentObject()->VertexPool[ VertexIndices[0] ] + (dir * dist);
 	return GetParentObject()->GetActualBrush()->ActorToWorld().TransformPosition( loc );
